requirements.txt
fastapi>=0.111,<0.120
uvicorn>=0.29,<0.35
pydantic>=2.0,<3.0
pyyaml>=6.0,<7.0

numpy>=1.26,<2.0
pandas>=2.0,<3.0
scikit-learn>=1.4,<2.0
scipy>=1.13,<2.0
joblib>=1.3,<2.0
xgboost==2.0.3

prometheus-client>=0.20,<0.22

httpx>=0.27,<0.29
requests>=2.31,<3.0
